Reproduce issue with focus events in slotted elements which appear under a dialog
Elements which are slotted into a dialog, though presented as though within a dialog may not have a parent which is a child of the dialog. This causes an issue with focus event handling in dialog-polyfill as the attempt to check if the element which is the subject of a focus event in 'safeBlur_' is a child of a dialog will jump to a parent element outside of the dialog. This reproduces this issue.